refactor: 레포지토리 ID 기반의 중첩 라우팅 구조로 개편
